@@ -15,41 +15,97 @@
   "homepage": "https://nats.io",
   "repository": {
     "type": "git",
-    "url": "git@github.com:nats-io/node-nats.git"
+    "url": "git@github.com:nats-io/nats.js.git"
   },
   "bugs": {
-    "url": "https://github.com/nats-io/node-nats/issues"
+    "url": "https://github.com/nats-io/nats.js/issues"
   },
-  "license": "MIT",
+  "license": "Apache-2.0",
   "private": false,
   "author": {
-    "name": "Derek Collison",
-    "email": "derek@apcera.com"
+    "name": "The NATS Authors"
   },
   "contributors": [],
-  "main": "./index.js",
+  "main": "index.js",
+  "types": "index.d.ts",
+  "files": [
+    "lib/",
+    "examples/",
+    "OWNERS.md",
+    "CODE-OF-CONDUCT.md",
+    "LICENSE",
+    "MAINTAINERS.md",
+    "index.d.ts"
+  ],
   "scripts": {
-    "lint": "jshint --reporter node_modules/jshint-stylish lib/*.js test/*.js test/support/*.js examples/*.js",
-    "depcheck": "dependency-check . lib/*",
-    "depcheck:unused": "dependency-check ./package.json --unused --no-dev lib/*",
-    "test:unit": "mkdir -p reports/ && NODE_ENV=test multi='spec=- xunit=reports/mocha-xunit.xml' istanbul cover _mocha -- -R mocha-multi --timeout 10000 --slow 750 && istanbul check-coverage",
-    "test": "npm run depcheck && npm run depcheck:unused && npm run lint && npm run test:unit",
-    "coveralls": "cat ./reports/coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
+    "build": "tsc",
+    "cjs": "deno run --allow-all bin/cjs-fix-imports.ts -o nats-base-client/ ./.deps/nats.deno/nats-base-client/",
+    "clean": "shx rm -Rf ./lib/* ./nats-base-client ./.deps",
+    "clone-nbc": "shx mkdir -p ./.deps && cd ./.deps && git clone --branch v1.1.0 https://github.com/nats-io/nats.deno.git",
+    "fmt": "deno fmt ./src/ ./examples/ ./test/",
+    "prepack": "npm run clone-nbc && npm run cjs && npm run check-package && npm run build",
+    "ava": "nyc ava --verbose -T 60000",
+    "test": "npm run check-package && npm run build && npm run ava",
+    "debug-test": "node node_modules/.bin/ava --verbose -T 6500000 --match",
+    "setup": "curl -fsSL https://deno.land/x/install/install.sh | sh",
+    "setup_win": "choco install deno",
+    "stage": "npm run clean && npm run clone-nbc && npm run cjs && shx rm -Rf ./deps/ && npm run build",
+    "cover:html": "nyc report --reporter=html && open coverage/index.html",
+    "coveralls": "shx mkdir -p ./coverage && nyc report --reporter=text-lcov > ./coverage/lcov.info",
+    "check-package": "deno run --allow-all bin/check-bundle-version.ts"
   },
   "engines": {
-    "node": ">= 0.10.x"
+    "node": ">= 10.0.0"
   },
-  "dependencies": {},
+  "dependencies": {
+    "nkeys.js": "^1.0.0-9"
+  },
   "devDependencies": {
-    "jshint": "2.9.x",
-    "jshint-stylish": "2.2.x",
-    "mocha": "2.5.x",
-    "mocha-multi": "0.9.x",
-    "mocha-lcov-reporter": "1.2.x",
-    "dependency-check": "2.5.x",
-    "istanbul": "0.4.x",
-    "coveralls": "^2.11.2",
-    "should": ">= 9.0.0"
+    "@types/node": "^14.14.34",
+    "ava": "^3.15.0",
+    "minimist": "^1.2.5",
+    "nyc": "^15.1.0",
+    "shx": "^0.3.3",
+    "tslint": "^6.1.3",
+    "typescript": "^4.1.3"
   },
-  "typings": "./index.d.ts"
+  "typings": "./index.d.ts",
+  "ava": {
+    "failFast": false,
+    "files": [
+      "./test/**/*.js",
+      "!./test/helpers/**/*.js"
+    ]
+  },
+  "nyc": {
+    "extension": [
+      ".ts",
+      ".js"
+    ],
+    "include": [
+      "src/**/*.ts",
+      "lib/src/**/*.js",
+      "nats-base-client/**/*.ts",
+      "lib/nats-base-client/**/*.js"
+    ],
+    "exclude": [
+      "nats-base-client/bench.ts",
+      "nats-base-client/codec.ts",
+      "nats-base-client/databuffer.ts",
+      "nats-base-client/denobuffer.ts",
+      "nats-base-client/muxsubscription.ts",
+      "nats-base-client/nkeys.ts",
+      "nats-base-client/nuid.ts",
+      "nats-base-client/parser.ts",
+      "nats-base-client/queued_iterator.ts",
+      "nats-base-client/servers.ts",
+      "nats-base-client/transport.ts",
+      "nats-base-client/util.ts",
+      "lib/test/**",
+      "examples/**",
+      "benchmark/**"
+    ],
+    "sourceMap": true,
+    "all": true
+  }
 }
