@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2020 The NATS Authors
+ * Copyright 2020 The NATS Authors
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -12,6 +12,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-"use strict";
-
-module.exports = require("./lib/src/mod");
+export * from "../nats-base-client/internal_mod";
