@@ -1,184 +1,213 @@
-/* jslint node: true */
-/* global describe: false, before: false, after: false, it: false */
-'use strict';
+/*
+ * Copyright 2018-2021 The NATS Authors
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+const test = require("ava");
+const { connect, ErrorCode, createInbox, Empty } = require(
+  "../",
+);
+const { Lock } = require("./helpers/lock");
 
-var NATS = require ('../'),
-    nsc = require('./support/nats_server_control'),
-    should = require('should');
+const u = "demo.nats.io:4222";
 
-describe('Max responses and Auto-unsub', function() {
+test("autounsub - max option", async (t) => {
+  const nc = await connect({ servers: u });
+  const subj = createInbox();
+  const sub = nc.subscribe(subj, { max: 10 });
+  for (let i = 0; i < 20; i++) {
+    nc.publish(subj);
+  }
+  await nc.flush();
+  t.is(sub.getReceived(), 10);
+  await nc.close();
+});
 
-  var PORT = 1422;
-  var server;
+test("autounsub - unsubscribe", async (t) => {
+  const nc = await connect({ servers: u });
+  const subj = createInbox();
+  const sub = nc.subscribe(subj, { max: 10 });
+  sub.unsubscribe(11);
+  for (let i = 0; i < 20; i++) {
+    nc.publish(subj);
+  }
+  await nc.flush();
+  t.is(sub.getReceived(), 11);
+  await nc.close();
+});
 
-  // Start up our own nats-server
-  before(function(done) {
-    server = nsc.start_server(PORT, done);
-  });
+test("autounsub - can unsub from auto-unsubscribed", async (t) => {
+  const nc = await connect({ servers: u });
+  const subj = createInbox();
+  const sub = nc.subscribe(subj, { max: 1 });
+  for (let i = 0; i < 20; i++) {
+    nc.publish(subj);
+  }
+  await nc.flush();
+  t.is(sub.getReceived(), 1);
+  sub.unsubscribe();
+  await nc.close();
+});
 
-  // Shutdown our server after we are done
-  after(function(){
-    server.kill();
-  });
-
-  it('should only received max responses requested', function(done) {
-    var nc = NATS.connect(PORT);
-    var WANT = 10;
-    var SEND = 20;
-    var received = 0;
-
-    nc.subscribe('foo', {'max':WANT}, function() {
-      received += 1;
-    });
-    for (var i=0; i<SEND; i++) {
-      nc.publish('foo');
+test("autounsub - can break to unsub", async (t) => {
+  const nc = await connect({ servers: u });
+  const subj = createInbox();
+  const sub = nc.subscribe(subj, { max: 20 });
+  const iter = (async () => {
+    for await (const m of sub) {
+      break;
     }
-    nc.flush(function() {
-      should.exists(received);
-      received.should.equal(WANT);
-      nc.close();
-      done();
-    });
-  });
+  })();
+  for (let i = 0; i < 20; i++) {
+    nc.publish(subj);
+  }
+  await nc.flush();
+  await iter;
+  t.is(sub.getProcessed(), 1);
+  await nc.close();
+});
 
-  it('should only received max responses requested (client support)', function(done) {
-    var nc = NATS.connect(PORT);
-    var WANT = 10;
-    var SEND = 20;
-    var received = 0;
+test("autounsub - can change auto-unsub to a higher value", async (t) => {
+  const nc = await connect({ servers: u });
+  const subj = createInbox();
+  const sub = nc.subscribe(subj, { max: 1 });
+  sub.unsubscribe(10);
+  for (let i = 0; i < 20; i++) {
+    nc.publish(subj);
+  }
+  await nc.flush();
+  t.is(sub.getReceived(), 10);
+  await nc.close();
+});
 
-    var sid = nc.subscribe('foo', function() {
-      received += 1;
-    });
-    for (var i=0; i<SEND; i++) {
-      nc.publish('foo');
+test("autounsub - request receives expected count with multiple helpers", async (
+  t,
+) => {
+  const nc = await connect({ servers: u });
+  const subj = createInbox();
+
+  const fn = (async (sub) => {
+    for await (const m of sub) {
+      m.respond();
     }
-    nc.unsubscribe(sid, WANT);
-
-    nc.flush(function() {
-      should.exists(received);
-      received.should.equal(WANT);
-      nc.close();
-      done();
-    });
   });
+  const subs = [];
+  for (let i = 0; i < 5; i++) {
+    const sub = nc.subscribe(subj);
+    const _ = fn(sub);
+    subs.push(sub);
+  }
+  await nc.request(subj);
+  await nc.drain();
 
-  it('should not complain when unsubscribing an auto-unsubscribed sid', function(done) {
-    var nc = NATS.connect(PORT);
-    var SEND = 20;
-    var received = 0;
+  let counts = subs.map((s) => {
+    return s.getReceived();
+  });
+  const count = counts.reduce((a, v) => a + v);
+  t.is(count, 5);
+});
 
-    var sid = nc.subscribe('foo', {'max':1}, function() {
-      received += 1;
-    });
-    for (var i=0; i<SEND; i++) {
-      nc.publish('foo');
+test("autounsub - manual request receives expected count with multiple helpers", async (
+  t,
+) => {
+  const nc = await connect({ servers: u });
+  const subj = createInbox();
+  const lock = Lock(5);
+
+  const fn = (async (sub) => {
+    for await (const m of sub) {
+      m.respond();
+      lock.unlock();
     }
-
-    nc.flush(function() {
-      nc.unsubscribe(sid);
-      should.exists(received);
-      received.should.equal(1);
-      nc.close();
-      done();
-    });
   });
+  for (let i = 0; i < 5; i++) {
+    const sub = nc.subscribe(subj);
+    const _ = fn(sub);
+  }
+  const replySubj = createInbox();
+  const sub = nc.subscribe(replySubj);
+  nc.publish(subj, Empty, { reply: replySubj });
+  await lock;
+  await nc.drain();
+  t.is(sub.getReceived(), 5);
+});
 
-  it('should allow proper override to a lesser value ', function(done) {
-    var nc = NATS.connect(PORT);
-    var SEND = 20;
-    var received = 0;
+test("autounsub - check subscription leaks", async (t) => {
+  let nc = await connect({ servers: u });
+  let subj = createInbox();
+  let sub = nc.subscribe(subj);
+  sub.unsubscribe();
+  t.is(nc.protocol.subscriptions.size(), 0);
+  await nc.close();
+});
 
-    var sid = nc.subscribe('foo', function() {
-      received += 1;
-      nc.unsubscribe(sid, 1);
-    });
-    nc.unsubscribe(sid, SEND);
+test("autounsub - check request leaks", async (t) => {
+  let nc = await connect({ servers: u });
+  let subj = createInbox();
 
-    for (var i=0; i<SEND; i++) {
-      nc.publish('foo');
+  // should have no subscriptions
+  t.is(nc.protocol.subscriptions.size(), 0);
+
+  let sub = nc.subscribe(subj);
+  const _ = (async () => {
+    for await (const m of sub) {
+      m.respond();
     }
+  })();
 
-    nc.flush(function() {
-      should.exists(received);
-      received.should.equal(1);
-      nc.close();
-      done();
-    });
+  // should have one subscription
+  t.is(nc.protocol.subscriptions.size(), 1);
+
+  let msgs = [];
+  msgs.push(nc.request(subj));
+  msgs.push(nc.request(subj));
+
+  // should have 2 mux subscriptions, and 2 subscriptions
+  t.is(nc.protocol.subscriptions.size(), 2);
+  t.is(nc.protocol.muxSubscriptions.size(), 2);
+
+  await Promise.all(msgs);
+
+  // mux subs should have pruned
+  t.is(nc.protocol.muxSubscriptions.size(), 0);
+
+  sub.unsubscribe();
+  t.is(nc.protocol.subscriptions.size(), 1);
+  await nc.close();
+});
+
+test("autounsub - check cancelled request leaks", async (t) => {
+  let nc = await connect({ servers: u });
+  let subj = createInbox();
+
+  // should have no subscriptions
+  t.is(nc.protocol.subscriptions.size(), 0);
+
+  let rp = nc.request(subj, Empty, { timeout: 100 });
+
+  t.is(nc.protocol.subscriptions.size(), 1);
+  t.is(nc.protocol.muxSubscriptions.size(), 1);
+
+  // the rejection should be timeout
+  const lock = Lock();
+  rp.catch((rej) => {
+    t.true(
+      rej.code === ErrorCode.Timeout || rej.code === ErrorCode.NoResponders,
+    );
+    lock.unlock();
   });
 
-  it('should allow proper override to a higher value', function(done) {
-    var nc = NATS.connect(PORT);
-    var WANT = 10;
-    var SEND = 20;
-    var received = 0;
-
-    var sid = nc.subscribe('foo', function() {
-      received += 1;
-    });
-    nc.unsubscribe(sid, 1);
-    nc.unsubscribe(sid, WANT);
-
-    for (var i=0; i<SEND; i++) {
-      nc.publish('foo');
-    }
-
-    nc.flush(function() {
-      should.exists(received);
-      received.should.equal(WANT);
-      nc.close();
-      done();
-    });
-  });
-
-  it('should only receive N msgs in request mode with multiple helpers', function(done) {
-    /* jshint loopfunc: true */
-    var nc = NATS.connect(PORT);
-    var received = 0;
-
-    // Create 5 helpers
-    for (var i=0; i<5; i++) {
-      nc.subscribe('help', function(msg, reply) {
-        nc.publish(reply, 'I can help!');
-      });
-    }
-
-    nc.request('help', null, {'max':1}, function() {
-      received += 1;
-      nc.flush(function() {
-        should.exists(received);
-        received.should.equal(1);
-        nc.close();
-        done();
-      });
-    });
-
-  });
-
-  it('should not leak subscriptions when using max', function(done) {
-    /* jshint loopfunc: true */
-    var nc = NATS.connect(PORT);
-    var received = 0;
-
-    nc.subscribe('help', function(msg, reply) {
-      nc.publish(reply, 'I can help!');
-    });
-
-    // Create 5 requests
-    for (var i=0; i<5; i++) {
-      nc.request('help', null, {'max':1}, function() {
-	received += 1;
-      });
-    }
-    nc.flush(function() {
-      setTimeout(function() {
-	received.should.equal(5);
-	Object.keys(nc.subs).length.should.equal(1);
-	nc.close();
-	done();
-      }, 100);
-    });
-
-  });
-
+  await lock;
+  // mux subs should have pruned
+  t.is(nc.protocol.muxSubscriptions.size(), 0);
+  await nc.close();
 });
