@@ -1,302 +1,250 @@
-/* jslint node: true */
-/* global describe: false, before: false, after: false, it: false */
-'use strict';
+/*
+ * Copyright 2018-2021 The NATS Authors
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+const test = require("ava");
+const {
+  connect,
+} = require("../");
+const { NatsServer } = require("./helpers/launcher");
+const {
+  createInbox,
+  Events,
+  ErrorCode,
+  deferred,
+  DebugEvents,
+} = require("../lib/nats-base-client/internal_mod");
+const { Lock } = require("./helpers/lock");
 
-var NATS = require ('../'),
-    nsc = require('./support/nats_server_control'),
-    should = require('should');
+test("reconnect - should receive when some servers are invalid", async (t) => {
+  const lock = Lock(1);
+  const servers = ["127.0.0.1:7", "demo.nats.io:4222"];
+  const nc = await connect({ servers: servers, noRandomize: true });
+  const subj = createInbox();
+  await nc.subscribe(subj, {
+    callback: () => {
+      lock.unlock();
+    },
+  });
+  nc.publish(subj);
+  await lock;
+  await nc.close();
+  // @ts-ignore
+  const a = nc.protocol.servers.getServers();
+  t.is(a.length, 1);
+  t.true(a[0].didConnect);
+});
 
-describe('Reconnect functionality', function() {
+test("reconnect - events", async (t) => {
+  const srv = await NatsServer.start();
 
-  var PORT = 1426;
-  var WAIT = 20;
-  var ATTEMPTS = 4;
-  var server;
-
-  // Start up our own nats-server
-  before(function(done) {
-    server = nsc.start_server(PORT, done);
+  let nc = await connect({
+    port: srv.port,
+    waitOnFirstConnect: true,
+    reconnectTimeWait: 100,
+    maxReconnectAttempts: 10,
   });
 
-  // Shutdown our server after we are done
-  after(function() {
-    server.kill();
-  });
+  let disconnects = 0;
+  let reconnecting = 0;
 
-  it('should not emit a reconnecting event if suppressed', function(done) {
-    var nc = NATS.connect({'port':PORT, 'reconnect':false});
-    should.exist(nc);
-    nc.on('connect', function() {
-      server.kill();
-    });
-    nc.on('reconnecting', function(/*client*/) {
-      done(new Error('Reconnecting improperly called'));
-    });
-    nc.on('close', function() {
-      nc.close();
-      server = nsc.start_server(PORT, done);
-    });
-  });
-
-  it('should emit a disconnect and a reconnecting event after proper delay', function(done) {
-    var nc = NATS.connect({'port':PORT, 'reconnectTimeWait':WAIT});
-    var startTime;
-    should.exist(nc);
-    nc.on('connect', function() {
-      server.kill();
-      startTime = new Date();
-    });
-    nc.on('reconnecting', function(/*client*/) {
-      var elapsed = new Date() - startTime;
-      elapsed.should.be.within(WAIT, 5*WAIT);
-      nc.close();
-      server = nsc.start_server(PORT, done);
-    });
-    nc.on('disconnect', function() {
-      var elapsed = new Date() - startTime;
-      elapsed.should.be.within(0, 5*WAIT);
-    });
-  });
-
-  it('should emit multiple reconnecting events and fail after maxReconnectAttempts', function(done) {
-    var nc = NATS.connect({'port':PORT, 'reconnectTimeWait':WAIT, 'maxReconnectAttempts':ATTEMPTS});
-    var startTime;
-    var numAttempts = 0;
-    nc.on('connect', function() {
-      server.kill();
-      startTime = new Date();
-    });
-    nc.on('reconnecting', function(/*client*/) {
-      var elapsed = new Date() - startTime;
-      elapsed.should.be.within(WAIT, 5*WAIT);
-      startTime = new Date();
-      numAttempts += 1;
-    });
-    nc.on('close', function() {
-      numAttempts.should.equal(ATTEMPTS);
-      nc.close();
-      server = nsc.start_server(PORT, done);
-    });
-  });
-
-  it('should emit reconnecting events indefinitely if maxReconnectAttempts is set to -1', function(done) {
-
-    var nc = NATS.connect({'port':PORT, 'reconnectTimeWait':WAIT, 'maxReconnectAttempts':-1});
-    var numAttempts = 0;
-
-    // stop trying after an arbitrary amount of elapsed time
-    var timeout = setTimeout(function() {
-      // restart server and make sure next flush works ok
-      if (server === null) {
-        server = nsc.start_server(PORT);
+  const status = nc.status();
+  (async () => {
+    for await (const e of status) {
+      switch (e.type) {
+        case "disconnect":
+          disconnects++;
+          break;
+        case "reconnecting":
+          reconnecting++;
+          break;
+        default:
+          t.log(e);
       }
-    }, 1000);
-
-    nc.on('connect', function() {
-      server.kill();
-      server = null;
-    });
-    nc.on('reconnecting', function(/*client*/) {
-      numAttempts += 1;
-      // attempt indefinitely to reconnect
-      nc.reconnects.should.equal(numAttempts);
-      nc.connected.should.equal(false);
-      nc.wasConnected.should.equal(true);
-      nc.reconnecting.should.equal(true);
-      // if maxReconnectAttempts is set to -1, the number of reconnects will always be greater
-      nc.reconnects.should.be.above(nc.options.maxReconnectAttempts);
-    });
-    nc.on('reconnect', function() {
-      nc.flush(function() {
-        nc.close();
-        done();
-      });
-    });
-  });
-
-  it('should succesfully reconnect to new server', function(done) {
-    var nc = NATS.connect({'port':PORT, 'reconnectTimeWait':100});
-    // Kill server after first successful contact
-    nc.flush(function() {
-      server.kill();
-      server = null;
-    });
-    nc.on('reconnecting', function(/*client*/) {
-      // restart server and make sure next flush works ok
-      if (server === null) {
-        server = nsc.start_server(PORT);
-      }
-    });
-    nc.on('reconnect', function() {
-      nc.flush(function() {
-        nc.close();
-        done();
-      });
-    });
-  });
-
-  it('should succesfully reconnect to new server with subscriptions', function(done) {
-    var nc = NATS.connect({'port':PORT, 'reconnectTimeWait':100});
-    // Kill server after first successful contact
-    nc.flush(function() {
-      server.kill();
-      server = null;
-    });
-    nc.subscribe('foo', function() {
-      nc.close();
-      done();
-    });
-    nc.on('reconnecting', function(/*client*/) {
-      // restart server and make sure next flush works ok
-      if (server === null) {
-        server = nsc.start_server(PORT);
-      }
-    });
-    nc.on('reconnect', function() {
-      nc.publish('foo');
-    });
-  });
-
-  it('should succesfully reconnect to new server with queue subscriptions correctly', function(done) {
-    var nc = NATS.connect({'port':PORT, 'reconnectTimeWait':100});
-    // Kill server after first successful contact
-    nc.flush(function() {
-      server.kill();
-      server = null;
-    });
-    var received = 0;
-    // Multiple subscribers
-    var cb = function cb() { received += 1; };
-    for (var i=0; i<5; i++) {
-      nc.subscribe('foo', {'queue':'myReconnectQueue'}, cb);
     }
-    nc.on('reconnecting', function(/*client*/) {
-      // restart server and make sure next flush works ok
-      if (server === null) {
-        server = nsc.start_server(PORT);
+  })().then();
+  await srv.stop();
+  try {
+    await nc.closed();
+  } catch (err) {
+    t.is(err.code, ErrorCode.ConnectionRefused);
+  }
+  t.is(disconnects, 1, "disconnects");
+  t.is(reconnecting, 10, "reconnecting");
+});
+
+test("reconnect - reconnect not emitted if suppressed", async (t) => {
+  const srv = await NatsServer.start();
+  let nc = await connect({
+    port: srv.port,
+    reconnect: false,
+  });
+
+  let disconnects = 0;
+  (async () => {
+    for await (const e of nc.status()) {
+      switch (e.type) {
+        case Events.Disconnect:
+          disconnects++;
+          break;
+        case DebugEvents.Reconnecting:
+          t.fail("shouldn't have emitted reconnecting");
+          break;
       }
-    });
-    nc.on('reconnect', function() {
-      nc.publish('foo', function() {
-	received.should.equal(1);
-	nc.close();
-	done();
-      });
-    });
+    }
+  })().then();
+
+  await srv.stop();
+  await nc.closed();
+  t.pass();
+});
+
+test("reconnect - reconnecting after proper delay", async (t) => {
+  const srv = await NatsServer.start();
+  let nc = await connect({
+    port: srv.port,
+    reconnectTimeWait: 500,
+    maxReconnectAttempts: 1,
   });
+  // @ts-ignore
+  const serverLastConnect = nc.protocol.servers.getCurrentServer().lastConnect;
 
-  it('should properly resync with inbound buffer non-nil', function(done) {
-    var nc = NATS.connect({'port':PORT, 'reconnectTimeWait':100});
-
-    // Send lots of data to ourselves
-    nc.on('connect', function() {
-      var sid = nc.subscribe('foo', function() {
-	// Kill server on first message, inbound should still be full.
-	server.kill();
-	nc.unsubscribe(sid);
-        server = nsc.start_server(PORT);
-      });
-      var b = new Buffer(4096).toString();
-      for (var i=0; i<1000; i++) {
-        nc.publish('foo', b);
+  const dt = deferred();
+  const _ = (async () => {
+    for await (const e of nc.status()) {
+      switch (e.type) {
+        case DebugEvents.Reconnecting:
+          const elapsed = Date.now() - serverLastConnect;
+          dt.resolve(elapsed);
+          break;
       }
-    });
+    }
+  })();
+  await srv.stop();
+  const elapsed = await dt;
+  t.true(elapsed >= 500 && elapsed <= 700, `elapsed was ${elapsed}`);
+  await nc.closed();
+});
 
-    nc.on('reconnect', function() {
-      nc.flush(function() {
-	nc.close();
-	done();
-      });
-    });
+test("reconnect - indefinite reconnects", async (t) => {
+  let srv = await NatsServer.start();
+
+  let nc = await connect({
+    port: srv.port,
+    reconnectTimeWait: 100,
+    maxReconnectAttempts: -1,
   });
 
-  it('should not crash when sending a publish with a callback after connection loss', function(done) {
-    var nc = NATS.connect({'port':PORT, 'reconnectTimeWait':WAIT});
-    var startTime;
-    should.exist(nc);
-    nc.on('connect', function() {
-      server.kill();
-      startTime = new Date();
-    });
-    nc.on('disconnect', function() {
-      nc.publish('foo', 'bar', 'reply', function() {
-         // fails to get here, but should not crash
-      });
-      server = nsc.start_server(PORT);
-    });
-    nc.on('reconnect', function() {
-      nc.flush(function() {
-	nc.close();
-	done();
-      });
-    });
-  });
-
-  it('should execute callbacks if published during reconnect', function(done) {
-    var nc = NATS.connect({'port':PORT, 'reconnectTimeWait':100});
-    nc.on('reconnecting', function(/*client*/) {
-      // restart server
-      if (server === null) {
-        nc.publish('foo', function() {
+  let disconnects = 0;
+  let reconnects = 0;
+  let reconnect = false;
+  (async (t) => {
+    for await (const e of nc.status()) {
+      switch (e.type) {
+        case Events.Disconnect:
+          disconnects++;
+          break;
+        case Events.Reconnect:
+          reconnect = true;
           nc.close();
-          done();
-        });
-        server = nsc.start_server(PORT);
+          break;
+        case DebugEvents.Reconnecting:
+          reconnects++;
+          break;
       }
-    });
-    nc.on('connect', function() {
-      var s = server;
-      server = null;
-      s.kill();
-    });
+    }
+  })().then();
+
+  await srv.stop();
+
+  const lock = Lock(1);
+  setTimeout(async (t) => {
+    srv = await srv.restart();
+    lock.unlock();
+  }, 1000);
+
+  await nc.closed();
+  await srv.stop();
+  await lock;
+  await srv.stop();
+  t.true(reconnects > 5);
+  t.true(reconnect);
+  t.is(disconnects, 1);
+});
+
+test("reconnect - jitter", async (t) => {
+  let srv = await NatsServer.start();
+
+  let called = false;
+  const h = () => {
+    called = true;
+    return 15;
+  };
+
+  let hasDefaultFn;
+  let dc = await connect({
+    port: srv.port,
+    reconnect: false,
+  });
+  hasDefaultFn = typeof dc.options.reconnectDelayHandler === "function";
+
+  let nc = await connect({
+    port: srv.port,
+    maxReconnectAttempts: 1,
+    reconnectDelayHandler: h,
   });
 
-  it('should not lose messages if published during reconnect', function(done) {
-    // This checks two things if the client publishes while reconnecting:
-    // 1) the message is published when the client reconnects
-    // 2) the client's subscriptions are synced before the message is published
-    var nc = NATS.connect({'port':PORT, 'reconnectTimeWait':100});
-    nc.subscribe('foo', function() {
-      nc.close();
-      done();
-    });
-    nc.on('reconnecting', function(/*client*/) {
-      // restart server
-      if (server === null) {
-        nc.publish('foo');
-        server = nsc.start_server(PORT);
-      }
-    });
-    nc.on('connect', function() {
-      var s = server;
-      server = null;
-      s.kill();
-    });
+  await srv.stop();
+  await nc.closed();
+  await dc.closed();
+  t.true(called);
+  t.true(hasDefaultFn);
+});
+
+test("reconnect - internal disconnect forces reconnect", async (t) => {
+  const srv = await NatsServer.start();
+  const nc = await connect({
+    port: srv.port,
+    reconnect: true,
+    reconnectTimeWait: 200,
   });
 
-  it('should emit reconnect before flush callbacks are called', function(done) {
-    var nc = NATS.connect({'port':PORT, 'reconnectTimeWait':100});
-    var reconnected = false;
-    nc.on('reconnecting', function() {
-      // restart server
-      if (server === null) {
-        nc.flush(function() {
-          nc.close();
-          if (!reconnected) {
-            done(new Error('Flush callback called before reconnect emitted'));
-          }
-          done();
-        });
-        server = nsc.start_server(PORT);
+  let stale = false;
+  let disconnect = false;
+  const lock = Lock();
+  (async () => {
+    for await (const e of nc.status()) {
+      switch (e.type) {
+        case DebugEvents.StaleConnection:
+          stale = true;
+          break;
+        case Events.Disconnect:
+          disconnect = true;
+          break;
+        case Events.Reconnect:
+          lock.unlock();
+          break;
       }
-    });
-    nc.on('reconnect', function() {
-      reconnected = true;
-    });
-    nc.on('connect', function() {
-      var s = server;
-      server = null;
-      s.kill();
-    });
-  });
+    }
+  })().then();
+
+  nc.protocol.disconnect();
+  await lock;
+
+  t.true(disconnect, "disconnect");
+  t.true(stale, "stale");
+
+  await nc.close();
+  await srv.stop();
 });
